backend/requirements.txt
fastapi
uvicorn
duckdb
pandas
pydantic
google-genai
python-dotenv
